backend/requirements.txt
# Backend API dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
httpx==0.26.0
python-multipart==0.0.6

# Include RAG dependencies
-r ../rag/requirements.txt
